refactor(tangle-dapp): Move network selector to shared ui lib
