+++ apps/tangle-dapp/components/NetworkSelectorButton/index.tsx
@@ -1,0 +1,32 @@
+'use client';
+
+import NetworkSelectorButtonComp from '@webb-tools/tangle-shared-ui/components/NetworkSelectorDropdown';
+import { usePathname } from 'next/navigation';
+import { type FC, useMemo } from 'react';
+
+import { PagePath } from '../../types';
+
+const NetworkSelectorButton: FC = () => {
+  const pathname = usePathname();
+
+  // Disable network switching when in Liquid Staking page,
+  // since it would have no effect there.
+  const isInLiquidStakingPage = pathname.startsWith(PagePath.LIQUID_STAKING);
+
+  const isInBridgePath = useMemo(
+    () => pathname.startsWith(PagePath.BRIDGE),
+    [pathname],
+  );
+
+  if (isInBridgePath) {
+    return null;
+  }
+
+  return (
+    <NetworkSelectorButtonComp
+      isNetworkSwitchDisabled={isInLiquidStakingPage}
+    />
+  );
+};
+
+export default NetworkSelectorButton;
