+++ libs/tangle-shared-ui/src/components/NetworkSelectorDropdown/CustomRpcEndpointInput.tsx
@@ -1,9 +1,7 @@
 import { Save, SaveWithBg } from '@webb-tools/icons';
-import useLocalStorage, {
-  LocalStorageKey,
-} from '@webb-tools/tangle-shared-ui/hooks/useLocalStorage';
 import { Input } from '@webb-tools/webb-ui-components';
 import { FC, useCallback, useEffect, useRef, useState } from 'react';
+import useLocalStorage, { LocalStorageKey } from '../../hooks/useLocalStorage';
 
 export type CustomRpcEndpointInputProps = {
   id: string;
