+++ libs/tangle-shared-ui/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/tangle-shared-ui';
+export {};
