--- libs/tangle-shared-ui/src/lib/tangle-shared-ui.module.scss
@@ -1,7 +1,0 @@
-/*
- * Replace this with your own classes
- *
- * e.g.
- * .container {
- * }
-*/
