--- libs/tangle-shared-ui/src/lib/tangle-shared-ui.tsx
@@ -1,11 +1,0 @@
-import styles from './tangle-shared-ui.module.scss';
-
-export function TangleSharedUi() {
-  return (
-    <div className={styles['container']}>
-      <h1>Welcome to TangleSharedUi!</h1>
-    </div>
-  );
-}
-
-export default TangleSharedUi;
